Revert "Extending the ordered-changestream-buffer-to-sourcedb to include Post…"
This reverts commit d505cf6d2744cb91873d3e8eda666a491bb23ab3.
