+++ v2/ordered-changestream-buffer-to-sourcedb/src/main/java/com/google/cloud/teleport/v2/templates/dao/DaoFactory.java
@@ -15,51 +15,25 @@
  */
 package com.google.cloud.teleport.v2.templates.dao;
 
-import com.google.cloud.teleport.v2.templates.common.ProcessingContext;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 /** DaoFactory, currently only supports MySql. */
 public class DaoFactory {
+  private String sqlUrl = "";
+  private String sqlUser = "";
+  private String sqlPasswd = "";
 
-  private static final Logger LOG = LoggerFactory.getLogger(DaoFactory.class);
+  private MySqlDao mySqlDao;
 
-  public static Dao getDao(ProcessingContext taskContext) throws IllegalArgumentException {
-    LOG.info(String.format("Getting the %s DAO", taskContext.getSourceDbType()));
-    if ("mysql".equals(taskContext.getSourceDbType())) {
-      String sqlUrl =
-          "jdbc:mysql://"
-              + taskContext.getShard().getHost()
-              + ":"
-              + taskContext.getShard().getPort()
-              + "/"
-              + taskContext.getShard().getDbName();
-      return new MySqlDao(
-          sqlUrl,
-          taskContext.getShard().getUserName(),
-          taskContext.getShard().getPassword(),
-          taskContext.getShard().getLogicalShardId(),
-          taskContext.getEnableSourceDbSsl(),
-          taskContext.getEnableSourceDbSslValidation(),
-          "");
-    } else if ("postgresql".equals(taskContext.getSourceDbType())) {
-      String sqlUrl =
-          "jdbc:postgresql://"
-              + taskContext.getShard().getHost()
-              + ":"
-              + taskContext.getShard().getPort()
-              + "/"
-              + taskContext.getShard().getDbName();
-      return new PostgreSqlDao(
-          sqlUrl,
-          taskContext.getShard().getUserName(),
-          taskContext.getShard().getPassword(),
-          taskContext.getShard().getLogicalShardId(),
-          taskContext.getEnableSourceDbSsl(),
-          taskContext.getEnableSourceDbSslValidation(),
-          "");
+  public DaoFactory(String sqlUrl, String sqlUser, String sqlPasswd) {
+    this.sqlUrl = sqlUrl;
+    this.sqlUser = sqlUser;
+    this.sqlPasswd = sqlPasswd;
+  }
+
+  public MySqlDao getMySqlDao(String shardId) {
+    if (this.mySqlDao != null) {
+      return this.mySqlDao;
     }
-    throw new IllegalArgumentException(
-        String.format("No DAO found for %s", taskContext.getSourceDbType()));
+    this.mySqlDao = new MySqlDao(this.sqlUrl, this.sqlUser, this.sqlPasswd, shardId);
+    return this.mySqlDao;
   }
 }
