--- v2/ordered-changestream-buffer-to-sourcedb/src/main/java/com/google/cloud/teleport/v2/templates/processing/dml/DMLGeneratorFactory.java
@@ -1,33 +1,0 @@
-/*
- * Copyright (C) 2024 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.google.cloud.teleport.v2.templates.processing.dml;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class DMLGeneratorFactory {
-  private static final Logger LOG = LoggerFactory.getLogger(DMLGeneratorFactory.class);
-
-  public static DMLGenerator getDMLGenerator(String sourceDbType) throws IllegalArgumentException {
-    LOG.info(String.format("Getting DML generator for %s", sourceDbType));
-    if ("mysql".equals(sourceDbType)) {
-      return new MySqlDMLGenerator();
-    } else if ("postgresql".equals(sourceDbType)) {
-      return new PostgreSQLDMLGenerator();
-    }
-    throw new IllegalArgumentException(String.format("No DMLGenerator found for %s", sourceDbType));
-  }
-}
