+++ v2/ordered-changestream-buffer-to-sourcedb/src/main/java/com/google/cloud/teleport/v2/templates/processing/dml/package-info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2024 Google Inc.
+ * Copyright (C) 2023 Google Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -14,5 +14,5 @@
  * the License.
  */
 
-/** Handles the DML processing. */
+/** Has logic for DML creation. */
 package com.google.cloud.teleport.v2.templates.processing.dml;
