+++ v2/ordered-changestream-buffer-to-sourcedb/src/main/java/com/google/cloud/teleport/v2/templates/processing/handler/InputRecordProcessor.java
@@ -17,10 +17,8 @@ package com.google.cloud.teleport.v2.templates.processing.handler;
 
 import com.google.cloud.Timestamp;
 import com.google.cloud.teleport.v2.spanner.migrations.schema.Schema;
-import com.google.cloud.teleport.v2.templates.dao.Dao;
+import com.google.cloud.teleport.v2.templates.dao.MySqlDao;
 import com.google.cloud.teleport.v2.templates.processing.dml.DMLGenerator;
-import com.google.cloud.teleport.v2.templates.processing.dml.DMLGeneratorFactory;
-import java.sql.SQLException;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
@@ -38,12 +36,12 @@ public class InputRecordProcessor {
   private static final Logger LOG = LoggerFactory.getLogger(InputRecordProcessor.class);
 
   public static void processRecords(
-      String sourceDbType,
       List<String> recordList,
       Schema schema,
-      Dao dao,
+      MySqlDao dao,
       String shardId,
       String sourceDbTimezoneOffset) {
+
     try {
       boolean capturedlagMetric = false;
       long replicationLag = 0L;
@@ -52,7 +50,6 @@ public class InputRecordProcessor {
       Distribution lagMetric = Metrics.distribution(shardId, "replicationLagInSeconds_" + shardId);
 
       List<String> dmlBatch = new ArrayList<>();
-      DMLGenerator dmlGenerator = DMLGeneratorFactory.getDMLGenerator(sourceDbType);
       for (String s : recordList) {
         List<String> parsedValues = parseRecord(s);
         String tableName = parsedValues.get(0);
@@ -65,7 +62,7 @@ public class InputRecordProcessor {
         JSONObject keysJson = new JSONObject(keysJsonStr);
 
         String dmlStatement =
-            dmlGenerator.getDMLStatement(
+            DMLGenerator.getDMLStatement(
                 modType, tableName, schema, newValuesJson, keysJson, sourceDbTimezoneOffset);
         if (!dmlStatement.isEmpty()) {
           dmlBatch.add(dmlStatement);
@@ -80,25 +77,25 @@ public class InputRecordProcessor {
           Instant commitTsInst = Timestamp.parseTimestamp(commitTs).toSqlTimestamp().toInstant();
           replicationLag = ChronoUnit.SECONDS.between(commitTsInst, instTime);
         }
-
-        Instant daoStartTime = Instant.now();
-        dao.batchWrite(dmlBatch);
-        Instant daoEndTime = Instant.now();
-        LOG.info(
-            "Shard "
-                + shardId
-                + String.format(": Write to %s for ", sourceDbType)
-                + recordList.size()
-                + " took : "
-                + ChronoUnit.MILLIS.between(daoStartTime, daoEndTime)
-                + " milliseconds ");
-
-        numRecProcessedMetric.inc(
-            recordList.size()); // update the number of records processed metric
-
-        lagMetric.update(replicationLag); // update the lag metric
       }
-    } catch (IllegalArgumentException | SQLException e) {
+
+      Instant daoStartTime = Instant.now();
+      dao.batchWrite(dmlBatch);
+      Instant daoEndTime = Instant.now();
+      LOG.info(
+          "Shard "
+              + shardId
+              + ": Write to mysql for "
+              + recordList.size()
+              + " took : "
+              + ChronoUnit.MILLIS.between(daoStartTime, daoEndTime)
+              + " milliseconds ");
+
+      numRecProcessedMetric.inc(recordList.size()); // update the number of records processed metric
+
+      lagMetric.update(replicationLag); // update the lag metric
+
+    } catch (Exception e) {
       throw new RuntimeException("Failed to process records: " + e.getMessage());
     }
   }
