+++ v2/ordered-changestream-buffer-to-sourcedb/src/test/java/com/google/cloud/teleport/v2/templates/utils/ProcessingContextGeneratorTest.java
@@ -27,23 +27,6 @@ import org.junit.runners.JUnit4;
 public final class ProcessingContextGeneratorTest {
 
   @Test
-  public void processingContextForKafkaPostgreSQL() {
-
-    Map<String, ProcessingContext> response =
-        ProcessingContextGenerator.getProcessingContextForKafka(
-            "src/test/resources/pgsqlShard.json",
-            "postgresql",
-            "src/test/resources/kafkaConnectionProfile.json",
-            "src/test/resources/allMatchSession.json",
-            "+00:00",
-            true,
-            true);
-
-    assertEquals(response.size(), 2);
-    assertEquals(response.get("shardC").getBufferType(), "kafka");
-  }
-
-  @Test
   public void processingContextForKafka() {
 
     Map<String, ProcessingContext> response =
@@ -52,29 +35,10 @@ public final class ProcessingContextGeneratorTest {
             "mysql",
             "src/test/resources/kafkaConnectionProfile.json",
             "src/test/resources/allMatchSession.json",
-            "+00:00",
-            true,
-            true);
+            "+00:00");
 
     assertEquals(response.size(), 2);
     assertEquals(response.get("shardA").getBufferType(), "kafka");
-  }
-
-  @Test
-  public void processingContextForPubSubPostgreSQL() {
-    Map<String, ProcessingContext> response =
-        ProcessingContextGenerator.getProcessingContextForPubSub(
-            "src/test/resources/pgsqlShard.json",
-            "postgresql",
-            "tada",
-            "src/test/resources/allMatchSession.json",
-            42,
-            "+00:00",
-            true,
-            true);
-
-    assertEquals(response.size(), 2);
-    assertEquals(response.get("shardC").getBufferType(), "pubsub");
   }
 
   @Test
@@ -86,9 +50,7 @@ public final class ProcessingContextGeneratorTest {
             "tada",
             "src/test/resources/allMatchSession.json",
             42,
-            "+00:00",
-            true,
-            true);
+            "+00:00");
 
     assertEquals(response.size(), 2);
     assertEquals(response.get("shardA").getBufferType(), "pubsub");
