--- v2/ordered-changestream-buffer-to-sourcedb/src/test/resources/pgsqlShard.json
@@ -1,18 +1,0 @@
-[
-  {
-  "logicalShardId":"shardC",
-  "host":"hostShardC",
-  "user":"test",
-  "password":"test",
-  "port":"5432",
-  "dbName":"test"
-  },
- {
-  "logicalShardId":"shardD",
-  "host":"hostShardD",
-  "user":"test",
-  "password":"test",
-  "port":"5432",
-  "dbName":"test"
-   }
-]
